Rename packages to unique names for Java 9+ modules sake
These changes rename the packages of this Hibernate dialect so they begin
with "com.github.gwenn.sqlite.dialect", which should be a pretty unique
choice.

The rationale behind this renaming is allowing Java 9+ automatic module
creation to work fine, because otherwise both Hibernate and this dialect
define the "org.hibernate.dialect" package, which is illegal when dealing
with Java 9+ projects that are not in the unnamed module. A workaround
is shading both Hibernate and this dialect in a single JAR, but that is
a mantainance nightmare.

Hibernate documentation doesn't mention that dialects must be in the
"org.hibernate.dialect" package, these changes work fine, and definitely
there are official dialects which are not in that package, like this one:
https://docs.jboss.org/hibernate/ogm/5.4/api/org/hibernate/ogm/datastore/mongodb/MongoDBDialect.html,
so I'm confident in them not breaking anything for other use cases
(unless they use reflection for some reason).